Refactor chatbot and RAG data repositories
- Deleted old chatbot configuration repository and RAG data repository files.
- Created new chatbot configuration repository with updated structure and methods.
- Implemented chat history repository for managing chat history entries.
- Added new RAG repository with enhanced functionality for handling RAG data.
- Updated import paths and organized repository structure for better maintainability.
