+++ src/commands/slash/chatbot/help.ts
@@ -2,7 +2,7 @@ import discord from "discord.js";
 
 import Formatter from "../../../utils/format";
 import { EmbedTemplate } from "../../../core/embed/template";
-import { ChatbotConfigRepository } from "../../../events/database/repo/chatbot_config";
+import { ChatbotConfigRepository } from "../../../events/database/repo/chat_bot";
 
 
 export const handleHelp = async (
