+++ src/commands/slash/chatbot/info.ts
@@ -1,7 +1,7 @@
 import discord from "discord.js";
 
 import { EmbedTemplate } from "../../../core/embed/template";
-import { ChatbotConfigRepository } from "../../../events/database/repo/chatbot_config";
+import { ChatbotConfigRepository } from "../../../events/database/repo/chat_bot";
 
 
 export const handleInfo = async (
