+++ src/commands/slash/chatbot/upload_rag.ts
@@ -3,7 +3,7 @@ import discord from "discord.js";
 import { RAG } from "../../../core/ai";
 import { Embedding } from "../../../core/ai";
 import { EmbedTemplate } from "../../../core/embed/template";
-import { RagRepository } from "../../../events/database/repo/rag_data";
+import { RagRepository } from "../../../events/database/repo/chat_bot";
 
 
 export const handleUploadRag = async (
