+++ src/events/database/repo/chat_bot/config.ts
@@ -1,7 +1,7 @@
 import { Repository, DataSource } from "typeorm";
 
-import client from "../../../salt";
-import { ChatbotConfig } from "../entities/chatbot_config";
+import client from "../../../../salt";
+import { ChatbotConfig } from "../../entities/chat_bot";
 
 
 export class ChatbotConfigRepository {
